fix: refactor go packages "grcp" and "valid" into "pkg"
All go services now only depend on "pgk", but not other services

The whitespace changes is a quickfix for a flaky test.
